requirements.txt
claude-agent-sdk
pyyaml>=6.0.0
pytest>=8.0.0
